fix : 엔티티 수정 - 엔티티 id를 int 대신 long으로 통일 - Student -> Users 테이블명 변경

+++ src/main/java/com/example/tiggle/controller/TestController.java
@@ -26,7 +26,7 @@ public class TestController {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
 //        CustomUserDetails userDetails = (CustomUserDetails) authentication.getPrincipal();
 
-        int userId = userDetails.getUserId();
+        Long userId = userDetails.getUserId();
         String encryptedUserKey = userDetails.getEncryptedUserKey();
 
         // 암호화된 userKey 복호화
